implemented main overview, deleted listoftrains and listofstops, small fixes
