+++ debug.log
@@ -34,3 +34,8 @@
 [1002/130221.786:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [1002/130221.786:ERROR:crash_report_database_win.cc(469)] failed to stat report
 [1002/130221.786:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1002/145145.018:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1002/145145.020:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1002/145145.020:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1002/145145.020:ERROR:crash_report_database_win.cc(469)] failed to stat report
+[1002/145145.020:ERROR:crash_report_database_win.cc(469)] failed to stat report
