+++ package.json
@@ -1,5 +1,4 @@
 {
-  "proxy": "http://localhost:5000/",
   "name": "trainfrontend",
   "version": "0.1.0",
   "private": true,
