--- src/Styles/ListofTrains.css
@@ -1,3 +1,0 @@
-/* .stopidTable {
-  display: none;
-} */
