+++ src/Styles/MainOverview.css
@@ -1,0 +1,3 @@
+.stopidTable {
+  display: none;
+}
